Split Coliquidity tests
